Support customising config path (#301)
In our company we're using this gem in monorepo - where there are
multiple Rails applications and Rails engines. This PR allows to specify
custom location for configuration file so it can be shared across
multiple applications.

Note: I couldn't reopen #210 - no privileges. So creating new PR instead
that is synced with main branch.

+++ lib/annotate_rb/config_finder.rb
@@ -7,6 +7,8 @@ module AnnotateRb
     DOTFILE = ".annotaterb.yml"
 
     class << self
+      attr_accessor :config_path
+
       def find_project_root
         # We should expect this method to be called from a Rails project root and returning it
         # e.g. "/Users/drwl/personal/annotaterb/dummyapp"
@@ -14,6 +16,8 @@ module AnnotateRb
       end
 
       def find_project_dotfile
+        return @config_path if @config_path && File.exist?(@config_path)
+
         [
           find_project_root.join(DOTFILE),
           find_project_root.join("config", DOTFILE.delete_prefix(".")),
